Refactor code and breakdown to smaller chunks for easier readability and referencing.
