+++ backend/src/apis/CDA/main.py
@@ -35,9 +35,27 @@ def aug_process():
 
     new_version = zip_check_dataset()
     time_print(start, "Zipping old datasets")
-    dataset_fols, template_dict, count_dict = get_dicts(image_fols, new_version)
+    dataset_fols, template_dict, count_dict = get_dicts(new_version, image_fols)
     time_print(start, "Copying Files and Templating all NGs")
     augmenting(g_path, file_names, template_dict, count_dict, dataset_fols, augment_set)
     time_print(start, "Augmenting all NGs")
     train_val_split(dataset_fols, image_fols, augment_set)
     time_print(start, "Train Val Split")
+
+
+def get_files(good):
+    files = {}
+    for fol in os.listdir(dire.image_path):
+        if good and fol.lower() == "g":
+            continue
+        for file in os.listdir(os.path.join(dire.image_path, fol)):
+            if file.split(".")[-1] not in ["jpg", "jpeg", "png", "tiff"]:
+                files[f"{fol}/{file}"] = os.listdir(
+                    os.path.join(dire.image_path, fol, file)
+                )
+            else:
+                if fol not in files.keys():
+                    files[fol] = []
+                files[fol].append(file)
+
+    return files
