+++ backend/src/core/json/augment.json
@@ -1,1 +1,1 @@
-{"target": 10000, "split": 20, "format": ["jpg", "jpeg", "png", "tiff"]}
+{ "target": 10000, "split": 20 }
